Configures Locate to use the new secrets package
This commit configures the Locate Service to use the new secrets
package, which fetches the JWT signer and verify keys from the Google
Secret Manager at runtime.

Additionally, this PR removes all of the old project-specific GAE config
files in favor of templates for both the locate and locate-platform
services.