Adds new package named secrets (#54)
* Adds new package named secrets

* Adds period to end of package description

* Returns nil instead of empty byte slice

* Removes unnecessary conditional and return in function

* Uses clearer, self-describing names in unit tests

The previous variable names were often unclear. For example, assigning
the value "no-errors" to a variable named "err". This commit cleans up
variable naming to make the logic clearer and somewhat self-describing.

* Makes success the default case in fakeIter tests

* Uses on a single wantErr var for fakeIter

* Fixes test Test_getSecretVersions()

Previously, fakeIter was returning an error even for a test case where
the intention was for fakeIter to succeed, and for the overall test to
fail because the single version returned by the function was one
disabled secret version.

* Uses table-driven tests for Test_getSecret()

Additionally, fixes the error message when the err value received
contradicts the value of wantErr.