Configures Locate to use the new secrets package (#56)
* Configures Locate to use the new secrets package

This commit configures the Locate Service to use the new secrets
package, which fetches the JWT signer and verify keys from the Google
Secret Manager at runtime.

Additionally, this PR removes all of the old project-specific GAE config
files in favor of templates for both the locate and locate-platform
services.

* Changes names of JWT flags and variables

The previous names were used for when the variables held the actual JWT
keys. Now they just point to the name of a secret in the SM. This
renames the flags and variables to more accurate represent the values
they hold.

* Groups one build step more logically

* Adds clarifying comment about cfg.LoadVerifier()