Increases max instances to 20 for Locate Service
There is no need to limit the number of instances of the Locate Service
that AppEngine can launch to scale up the app. The default max is 8, and
this change sets the max to what I hope will be a very generous ceiling
of 20. I removed the min number of instances configurations, since its
value (2) was equal to the default for automatic scaling.