Increases max instances to 20 for Locate Service (#57)
* Increases max instances to 20 for Locate Service

There is no need to limit the number of instances of the Locate Service
that AppEngine can launch to scale up the app. The default max is 8, and
this change sets the max to what I hope will be a very generous ceiling
of 20. I removed the min number of instances configurations, since its
value (2) was equal to the default for automatic scaling.

* Changes location of go-jose.v2 package

Builds were failing while fetching the package. It used to be located
at:

"gopkg.in/square/go-jose.v2/jwt"

... but is apparently now located at:

"gopkg.in/go-jose/go-jose.v2/jwt"

* Updates Travis build env to use Go 1.16

* Causes Travis to use Ubuntu Focal for builds

* Reinstalls ca-certificates Ubuntu package

* Moves reinstall of ca-certs to before_install

* Uses sudo to run apt reinstall ca-certificates

* Runs apt-update before apt-reinstall

* Revert "Changes location of go-jose.v2 package"

This reverts commit 8a99ffa1e9d0d8a5a3095ad901e0a876a8f86736.