Uses m-lab/go/host package to determine hostname
Virtual machines that are part of a managed instance group will have
hostnames/machinenames that have a random suffix, and many of mlab's
tools expect hostnames without this suffix. This commit uses the
hostname as parsed by the `host` package rather than the raw node Name
passed in through the environment. This should result in a valid
hostname that does not include any random suffixes.