Uses m-lab/go/host v0.1.58 to preserve service name
In the case of virtual nodes that are part of a managed instane group,
the node name and hostname will have a random 4 char suffix. This suffix
isn't wanted by heartbeat or Locate. This commit updates the included
version of m-lab/go to bring in changes to the host package, which adds
new functions for returning the hostname with or without service names
and suffixes.