Use updated version of m-lab/go/host, which supports service names (#117)
* Uses m-lab/go/host package to determine hostname

Virtual machines that are part of a managed instance group will have
hostnames/machinenames that have a random suffix, and many of mlab's
tools expect hostnames without this suffix. This commit uses the
hostname as parsed by the `host` package rather than the raw node Name
passed in through the environment. This should result in a valid
hostname that does not include any random suffixes.

* Uses m-lab/go/host v0.1.58 to preserve service name

In the case of virtual nodes that are part of a managed instane group,
the node name and hostname will have a random 4 char suffix. This suffix
isn't wanted by heartbeat or Locate. This commit updates the included
version of m-lab/go to bring in changes to the host package, which adds
new functions for returning the hostname with or without service names
and suffixes.

* Fixes unit test for location.go which uses  package