Reads value of var lbPath (/metadata/loadbalanced)
Previously, the file /metadata/loadbalanced only existed on MIG instances. We
recently decided that the file should exist on every machine with either a
"true" or "false" value. Locate was only checking for the file's existence to
determine if the machine was loadbalanced or not, but since this is no longer
a valid signal, this commit causes heartbeat to read the value of the file, and
only treat the machine as loadbalanced if the value is "true".