Reads value of var lbPath (/metadata/loadbalanced) (#191)
* Reads value of var lbPath (/metadata/loadbalanced)

Previously, the file /metadata/loadbalanced only existed on MIG instances. We
recently decided that the file should exist on every machine with either a
"true" or "false" value. Locate was only checking for the file's existence to
determine if the machine was loadbalanced or not, but since this is no longer
a valid signal, this commit causes heartbeat to read the value of the file, and
only treat the machine as loadbalanced if the value is "true".

* Fixes reading of var lbPath and associated error

Previously the code was calling os.ReadFile() and putting any error in lberr,
but rtx.Must() was inspecting variable err, which was always nil. This commit
has rtx.Must() properly look at var lberr.

Additionally, have the unit test write out a temporary file at lbPath and
remove it after the test.

* Don't fatally exit when /metadata/loadbalancer doesn't exist

Ultimately, we want to fatally exit, but for now the majority of machines on
the platform don't have this file.

* Updates a comment to reflect the current reality.

* Reverses conditional logic to be correct

We don't want to check for the value of the loadbalanced file until we are sure
that lberr==nil.