Adds a new "geo" format the /v2/siteinfo/registrations endpoint (#224)
* Adds a new Locate endpoint /v2/siteinfo/geo

The new endpoint will return a GeoJSON feature collection which represents all
machines that Locate knows about.

* Actually appends the Features to the FeatureCollection

Also, removes an incorrect change to handler.go, and updates a comment.

* Updates github.com/m-lab/go to v0.1.76

* Changes properties in GeoJSON returned by registrations?format=geo

The fields used by the status map on the M-Lab website are completely different
than what we have available to us from the Locate registrations. The status map
will have to start displaying different data with each measurement site.

* Fixes heartbeat unit tests

The unit tests were not expecting the "Org" field in the host.Name struct, but
that field is now populated in all V2 and V3 hostnames.