Factor filtering hosts into its own function for the registrations endpoint (#226)
* Considers only NDT registrations for siteinfo GeoJSON format

We don't want a separate GeoJSON feature for every experiment on every machine.
Instead, we only want a single feature for each machine. NDT runs on every
machine, so this should accurately represent the entire fleet. This is overly
NDT-centric, unfortunately, but is a reasonable start, I think.

* Factors filtering hosts into a separate function

Both the Hosts() and Geo() function use the filtering. I also renamed
Machines() to Hosts(), since this is more accurate as the returned values
experiment hostnames, not machine names.

* Sets appropriate headers for the registrations endpoint

Namely, we want a permissive CORS policy to allow any origin to make use of the
data provided by this endpoint.

* Refactors the filterHosts() function

The previous logic was not scalable and was difficult to read.

* Adds a machine property to the GeoJSON for registrations geo format

The machine format will be slightly different than the standart M-Lab machine
naming. Standard formatting is like <machine>-<site>, but this is going to be
<site>-<machine>, since that is the order that BYOS now uses.