fix: use PLATFORM_PROJECT for the auth service in openapi.yaml (#229)
* fix: split PROJECT and AUTH_PROJECT to handle the mlab-ns case

In the production environment, we have a single instance of the auth Cloud Run
service that lives in the mlab-oti project, but Locate is deployed to the
mlab-ns project instead. This means we need a separate variable to track the
project where Locate is deployed (PROJECT) vs the project where the auth
service lives (AUTH_PROJECT).

* use the already existing PLATFORM_PROJECT instead.